refactor: move sessions to users
