+++ backend/src/users/sessions/dto/updateSession.dto.ts
