+++ backend/src/users/status/status.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { StatusGateway } from './status.gateway';
 import { UsersService } from '../users.service';
-import { SessionsService } from '../../sessions/sessions.service';
+import { SessionsService } from '../sessions/sessions.service';
 import { SessionEntity, UserEntity } from '../../core/entities';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { S3ClientProvider } from '../../lib/aws/s3Client';
